+++ builder3Server/builder/pandoc.ts
@@ -14,7 +14,6 @@ export class Pandoc {
 
   async generate(input: PandocInput) {
     const cmd = this.createCommand(input);
-    console.log(cmd)
     const { stdout, stderr } = await exec(cmd);
     console.log("stdout:", stdout);
     console.log("stderr:", stderr);
@@ -22,17 +21,9 @@ export class Pandoc {
 
   createCommand(input: PandocInput): string {
     return `pandoc ${input.inputPath} -o ${input.outputPath} ${
-      input.isTableOfContents ? "--table-of-contents " : ""
+      input.isTableOfContents ? "--toc --toc-depth=1 " : ""
     } ${
-      input.metadataFile ? `--metadata-file=${input.metadataFile}` : ""
-    }  -V geometry:margin=1in  --highlight-style pygments  `;
+      input.metadataFile ? ` --metadata-file=${input.metadataFile}` : ""
+    }  -V geometry:margin=1in --highlight-style pygments  `;
   }
 }
-
-// const pandoc = new Pandoc();
-// pandoc.generate({
-//   inputPath: "static/prepared_all_algorithms.html",
-//   outputPath: "pdf/output_from_html_algorithms.pdf",
-//   isTableOfContents: true,
-//   metadataFile: "meta/handbook_algorithms.yaml",
-// });